Added overwrite protection when generating a year

+++ utils/generate_year.py
@@ -19,29 +19,34 @@ def generate_year(year: int) -> None:
     # Init file
     with open(new_folder + "__init__.py", 'w'):
         pass
-    # Task files
+    # Task files (OVERWRITE PROTECTED)
     for day in range(1, 26):
-        with open(template_folder + "task_template.py", 'r') as template_file:
-            template = template_file.read()
-        template = template.replace('8000', str(year))
-        template = template.replace('9999', str(day))
-        with open(new_folder + f"task_{day}.py", 'w') as task_file:
-            task_file.write(template)
+        # Create a new task file if it does not exist
+        if not os.path.exists(new_folder + f"task_{day}.py"):
+            with open(template_folder + "task_template.py", 'r') as template_file:
+                template = template_file.read()
+            template = template.replace('8000', str(year))
+            template = template.replace('9999', str(day))
+            with open(new_folder + f"task_{day}.py", 'w') as task_file:
+                task_file.write(template)
     # Tasks collection
     with open(template_folder + "tasks_template.py", 'r') as template_file:
         template = template_file.read()
     template = template.replace('9999', str(year))
     with open(new_folder + "tasks.py", 'w') as tasks:
         tasks.write(template)
-    # Answers JSON
-    shutil.copy2(template_folder + "answers.json", new_data_folder)
-    # txt files
+    # Answers JSON (OVERWRITE PROTECTED)
+    if not os.path.exists(template_folder + "answers.json"):
+        shutil.copy2(template_folder + "answers.json", new_data_folder)
+    # txt files (OVERWRITE PROTECTED)
     for day in range(1, 26):
         # TODO: Use API to automatically fill data
-        with open(new_data_folder + f"task{day}_data.txt", 'w'):
-            pass
-        with open(new_data_folder + f"task{day}_test.txt", 'w'):
-            pass
+        if not os.path.exists(new_data_folder + f"task{day}_data.txt"):
+            with open(new_data_folder + f"task{day}_data.txt", 'w'):
+                pass
+        if not os.path.exists(new_data_folder + f"task{day}_test.txt"):
+            with open(new_data_folder + f"task{day}_test.txt", 'w'):
+                pass
     print(f"Successfully created an AoC folder for: {year}")
